# Conflicts:
#	cas-server-support-pac4j/src/main/java/org/jasig/cas/support/pac4j/web/flow/ClientAction.java

+++ cas-server-support-oauth/src/main/java/org/jasig/cas/support/oauth/web/OAuth20AccessTokenController.java
@@ -24,8 +24,9 @@ import org.pac4j.core.util.CommonHelper;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.cloud.context.config.annotation.RefreshScope;
-import org.springframework.stereotype.Component;
+import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.servlet.ModelAndView;
 
 import javax.servlet.http.HttpServletRequest;
@@ -41,14 +42,14 @@ import java.util.Optional;
  * @since 3.5.0
  */
 @RefreshScope
-@Component("accessTokenController")
+@Controller("accessTokenController")
 public class OAuth20AccessTokenController extends BaseOAuthWrapperController {
 
     @Autowired
     @Qualifier("defaultRefreshTokenFactory")
     private RefreshTokenFactory refreshTokenFactory;
 
-    @RequestMapping(path=OAuthConstants.BASE_OAUTH20_URL + '/' + OAuthConstants.ACCESS_TOKEN_URL)
+    @RequestMapping(path = OAuthConstants.BASE_OAUTH20_URL + '/' + OAuthConstants.ACCESS_TOKEN_URL, method = RequestMethod.POST)
     @Override
     protected ModelAndView handleRequestInternal(final HttpServletRequest request, final HttpServletResponse response) throws Exception {
         response.setContentType("text/plain");
@@ -164,7 +165,7 @@ public class OAuth20AccessTokenController extends BaseOAuthWrapperController {
             }
             return null;
         }
-        if (token instanceof OAuthCode) {
+        if (token instanceof OAuthCode && !(token instanceof RefreshToken)) {
             this.ticketRegistry.deleteTicket(token.getId());
         }
